SSCSSI-316: Clean up Abstract Test Methods
